Whole project build system works
