--- common/index.js
@@ -1,7 +1,0 @@
-const util = require('./util');
-const settings = require('./settings');
-
-module.exports = {
-    ...util,
-    ...settings
-};
