--- common/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "common",
-  "version": "0.7.0",
-  "description": "Take Counter Common Components Library",
-  "main": "index.js",
-  "scripts": {
-    "lint": "eslint --max-warnings 0 .",
-    "test": "jest"
-  },
-  "author": "",
-  "license": "UNLICENSED",
-  "devDependencies": {
-    "@eslint/js": "^9.28.0",
-    "eslint": "^9.28.0",
-    "globals": "^16.2.0",
-    "jest": "^29.7.0"
-  }
-}
