+++ components/babel.config.cjs
@@ -1,0 +1,17 @@
+module.exports = {
+    presets: [
+      '@babel/preset-env',
+      ['@babel/preset-react', {runtime: 'automatic'}],
+    ],
+    plugins: [
+        [
+            'babel-plugin-transform-scss',
+            {
+                'include': [
+                    'node_modules',
+                    '../node_modules'
+                ] 
+            } 
+        ]
+    ]
+  };
