+++ components/jest.config.cjs
@@ -1,0 +1,10 @@
+/** @type {import('jest').Config} */
+
+const config = {
+    testEnvironment: 'jsdom',
+    setupFilesAfterEnv: [
+        './src/setupTest.js'
+    ]
+}
+
+module.exports = config;
