+++ components/src/Counter/TakeInputDisplay.jsx
