+++ components/src/Details/index.jsx
