+++ electron/src/main_window/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import TakeCounter from '../../components/src/TakeCounter';
+import TakeCounter from '../../../components/src/TakeCounter.jsx';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
