+++ electron/webpack.rules.cjs
@@ -35,7 +35,14 @@ module.exports = [
       // Translates CSS into CommonJS
       "css-loader",
       // Compiles Sass to CSS
-      "sass-loader",
+      {
+        loader: 'sass-loader',
+        options: {
+          sassOptions: {
+            quietDeps: true
+          }
+        }
+      }
     ],
   },
   // Put your webpack loader rules in this array.  This is where you would put
