+++ package.json
@@ -5,7 +5,6 @@
   "workspaces": [
     "electron",
     "web",
-    "common",
     "components"
   ],
   "scripts": {
@@ -13,9 +12,9 @@
     "web": "npm start -w web",
     "web:build": "npm run -w web build",
     "make": "npm run make -w electron",
-    "lint": "npm run lint --workspaces --if-present",
+    "lint": "npm run lint --workspaces",
     "test": "npm test --workspaces --if-present",
-    "test:ci": "npm run test:ci -w electron"
+    "test:ci": "npm run test:ci --workspaces --if-present"
   },
   "keywords": [],
   "author": {
